fix:修改部门id为Long

+++ youlai-admin/admin-api/src/main/java/com/youlai/admin/pojo/SysDept.java
@@ -13,7 +13,7 @@ public class SysDept extends BaseEntity {
 
     private String name;
 
-    private Integer parentId;
+    private Long parentId;
 
     private String treePath;
